render successful and components ready
