+++ src/components/body/error.jsx
@@ -1,0 +1,14 @@
+import React from 'react';
+
+function Error() {
+  return (
+    <div className='error-container'>
+      <p className='error'>
+        There was an error processing your request. Please try again without any
+        special characters!
+      </p>
+    </div>
+  );
+}
+
+export default Error;
