+++ src/components/body/loading.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+function loading() {
+  return (
+    <section className='loading-center'>
+      <div className='loading'></div>
+    </section>
+  );
+}
+
+export default loading;
